[FEATURE, Import] Überschreibe vorhandenes Schema nicht automatisch

+++ QGIS.sql
@@ -141,12 +141,15 @@ GRANT EXECUTE ON FUNCTION "QGIS"."imp_create_xp_gid"(character varying, characte
 COMMENT ON FUNCTION "QGIS"."imp_create_xp_gid"(character varying, character varying) IS 'Legt in der übergebenen Tabelle ein bigint-Feld xp_gid an; Funktion ist nötig, da der ALTER TABLE-Befehl aus QGIS selbst über QtSql nicht geht :-(';
 
 CREATE OR REPLACE FUNCTION "QGIS".imp_create_schema(
-    nspname character varying)
+    nspname character varying, ueberschreiben integer)
     RETURNS integer AS
 $BODY$
 BEGIN
-    EXECUTE 'DROP SCHEMA IF EXISTS ' ||
-        quote_ident(nspname) || ' CASCADE;';
+    IF ueberschreiben = 1 THEN
+        EXECUTE 'DROP SCHEMA IF EXISTS ' ||
+            quote_ident(nspname) || ' CASCADE;';
+    END IF;
+
     EXECUTE 'CREATE SCHEMA ' ||
         quote_ident(nspname) || ' ;';
     RETURN 1;
@@ -154,8 +157,8 @@ END;
 $BODY$
 LANGUAGE 'plpgsql' VOLATILE
 COST 100;
-GRANT EXECUTE ON FUNCTION "QGIS".imp_create_schema(character varying) TO xp_user;
-COMMENT ON FUNCTION "QGIS".imp_create_schema(character varying)
+GRANT EXECUTE ON FUNCTION "QGIS".imp_create_schema(character varying,integer) TO xp_user;
+COMMENT ON FUNCTION "QGIS".imp_create_schema(character varying,integer)
     IS 'Legt übergebenes Schema neu an; Funktion ist nötig, da der CREATE-Befehl aus QGIS selbst über QtSql nicht geht :-(';
 
 -- *****************************************************
